Automate setup of docker container to validate dependencies
The docker container currently is based on an LXDE based debian, which
doesn't match the tutorial, yet. But it was an approach supporting GUI
applications. It should be reworked to base on a plain ubuntu and add
the necessary GUI support on top.

If we don't create the project directory ourselves the COPY clause in
the Dockerfile creates it. But despite the current user being 'user' the
directory was owned by root so that the following pip3 commands would
fail. So instead create the directory in the "...-as-root.sh" script and
change ownership to the regular user.